refactor(challenge): ChallengeStatistics 체리 레벨 실시간 계산
